requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
elastic-transport==8.4.0
elasticsearch==7.17.6
memory-profiler==0.60.0
msgpack==1.0.4
psutil==5.9.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
reactivex==4.0.4
requests==2.28.1
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.12
websockets==10.3
